+++ core/lib/sn/Config.d.ts
@@ -8,7 +8,7 @@ export declare class Config implements IConfig {
     hPathFn2Exts: #IFn2Path;
     static readonly EXT_SPRITE = "png|jpg|jpeg|json|svg|webp|mp4|webm";
     static readonly EXT_SCRIPT = "sn|ssn";
-    static readonly EXT_FONT = "woff2|otf|ttf";
+    static readonly EXT_FONT = "woff2|woff|otf|ttf";
     static readonly EXT_SOUND = "mp3|m4a|ogg|aac|flac|wav";
     static readonly EXT_HTML = "htm|html";
     constructor(sys: SysBase);